+++ src/app/admin/admin.module.ts
@@ -41,6 +41,13 @@ import { AdvertiserListItemComponent } from 'admin/users/advertiser-list/adverti
 import { PublisherListComponent } from 'admin/users/publisher-list/publisher-list.component';
 import { PublisherListItemComponent } from 'admin/users/publisher-list/publisher-list-item/publisher-list-item.component';
 import { PanelPlaceholdersSettingsComponent } from 'admin/panel-placeholders-settings/panel-placeholders-settings.component';
+import { MatRadioModule } from '@angular/material/radio';
+import { SiteOptionsSettingsComponent } from './general-settings/site-options-settings/site-options-settings.component';
+import { BanUserDialogComponent } from './dialog/ban-user-dialog/ban-user-dialog.component';
+import { DeleteUserDialogComponent } from './dialog/delete-user-dialog/delete-user-dialog.component'
+import { MatDialogModule } from '@angular/material/dialog'
+import { MatMenuModule } from '@angular/material/menu'
+import { MatButtonModule } from '@angular/material/button'
 
 @NgModule({
   imports: [
@@ -62,6 +69,10 @@ import { PanelPlaceholdersSettingsComponent } from 'admin/panel-placeholders-set
     MatDatepickerModule,
     MatSelectModule,
     MatTabsModule,
+    MatRadioModule,
+    MatDialogModule,
+    MatMenuModule,
+    MatButtonModule,
   ],
   providers: [
     AdminGuard,
@@ -89,9 +100,14 @@ import { PanelPlaceholdersSettingsComponent } from 'admin/panel-placeholders-set
     RejectedDomainsComponent,
     PanelBlockadeComponent,
     PanelPlaceholdersSettingsComponent,
+    SiteOptionsSettingsComponent,
+    BanUserDialogComponent,
+    DeleteUserDialogComponent,
   ],
   entryComponents: [
     SuccessSnackbarComponent,
+    BanUserDialogComponent,
+    DeleteUserDialogComponent,
   ]
 })
 export class AdminModule {