+++ src/app/admin/users/user-list/user-list-item/user-list-item.component.ts
@@ -14,6 +14,14 @@ import { SessionService } from '../../../../session.service'
 import { CODE, CRYPTO } from 'common/utilities/consts'
 import { User } from 'models/user.model'
 import { ShowDialogOnError } from 'store/common/common.actions'
+import { MatDialog } from '@angular/material/dialog'
+import { BanUserDialogComponent } from 'admin/dialog/ban-user-dialog/ban-user-dialog.component'
+import { DeleteUserDialogComponent } from 'admin/dialog/delete-user-dialog/delete-user-dialog.component'
+import {
+  UserConfirmResponseDialogComponent
+} from 'common/dialog/user-confirm-response-dialog/user-confirm-response-dialog.component'
+import { BanUser, DeleteUser, UnbanUser } from 'store/admin/admin.actions'
+
 
 @Component({
   selector: 'app-user-list-item',
@@ -36,6 +44,7 @@ export class UserListItemComponent {
     private impersonationService: ImpersonationService,
     private sessionService: SessionService,
     private router: Router,
+    private dialog: MatDialog,
   ) {
     this.loggedUser = sessionService.getUser()
   }
@@ -139,6 +148,9 @@ export class UserListItemComponent {
     else if (this.user.isAgency) {
       return 'Agency'
     }
+    else if (this.user.isBanned) {
+      return 'Banned'
+    }
     else if (this.user.isAdvertiser && this.user.isPublisher) {
       return 'Adv / Pub'
     }
@@ -163,5 +175,44 @@ export class UserListItemComponent {
     else {
       return this.user.isAdvertiser ? 'Advertiser' : 'Publisher'
     }
+  }
+
+  showBanConfirmationDialog(){
+    this.dialog.open(BanUserDialogComponent, {
+      autoFocus: false,
+      data: this.user
+    })
+      .afterClosed()
+      .subscribe(result => {
+        if(result) {
+          this.store.dispatch(new BanUser({ id: this.user.id, reason: result }))
+        }
+      })
+  }
+
+  showUnbanConfirmationDialog(){
+    this.dialog.open(UserConfirmResponseDialogComponent, {
+      data: {
+        message: `Do you want unban user ${this.user.email}?`
+      }
+    })
+      .afterClosed()
+      .subscribe(result => {
+        if(result){
+          this.store.dispatch(new UnbanUser(this.user.id))
+        }
+      })
+  }
+
+  showDeleteConfirmationDialog(){
+    this.dialog.open(DeleteUserDialogComponent, {
+      data: this.user.email
+    })
+      .afterClosed()
+      .subscribe(result => {
+        if(result){
+          this.store.dispatch(new DeleteUser(this.user.id))
+        }
+      })
   }
 }