+++ src/app/advertiser/edit-campaign/edit-campaign-basic-info/edit-campaign-basic-information.component.ts
@@ -255,4 +255,8 @@ export class EditCampaignBasicInformationComponent extends HandleSubscription im
       })
     this.subscriptions.push(subscription)
   }
+
+  get isTaxonomy(): boolean {
+    return this.media.length > 0
+  }
 }