+++ src/app/advertiser/edit-campaign/edit-campaign-bid-strategy/edit-campaign-bid-strategy.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { MatDialog } from '@angular/material';
 import { ActivatedRoute, Router } from '@angular/router';
 import { Actions, ofType } from '@ngrx/effects';
-import { Store } from '@ngrx/store';
+import { Action, Store } from '@ngrx/store';
 import { first } from 'rxjs/operators';
 import { AppState } from 'models/app-state.model';
 import { BidStrategy, Campaign, CampaignsConfig } from 'models/campaign.model';
@@ -44,37 +44,27 @@ export class EditCampaignBidStrategyComponent extends HandleSubscription impleme
     super();
   }
 
-  ngOnInit() {
+  ngOnInit(): void {
     this.store.dispatch(new LoadCampaignsConfig());
     this.fetchBidStrategies();
   }
 
   fetchBidStrategies(): void {
-    this.bidStrategyService.getBidStrategies(true).subscribe(
+    this.campaign = this.route.snapshot.parent.data.campaign
+    this.bidStrategyService.getBidStrategies(this.campaign.basicInformation.medium, this.campaign.basicInformation.vendor, true).subscribe(
       (bidStrategies) => {
         this.bidStrategies = bidStrategies;
-        this.getCampaignFromStore();
+        const bidStrategyUuid = this.campaign.bidStrategy.uuid;
+        if (bidStrategyUuid && (-1 !== this.bidStrategies.findIndex((bidStrategy) => bidStrategy.uuid === bidStrategyUuid))) {
+          this.bidStrategyUuidSelected = bidStrategyUuid;
+        }
+        this.isLoading = false;
       },
       (error) => {
         const status = error.status ? error.status : 0;
         this.store.dispatch(new ShowDialogOnError(`Reload the page to load data. Error code (${status})`));
         this.isLoading = false;
       });
-  }
-
-  getCampaignFromStore(): void {
-    const subscription = this.store.select('state', 'advertiser', 'lastEditedCampaign')
-      .pipe(first())
-      .subscribe((lastEditedCampaign: Campaign) => {
-        this.campaign = lastEditedCampaign;
-        const bidStrategyUuid = this.campaign.bidStrategy.uuid;
-        if (bidStrategyUuid && (-1 !== this.bidStrategies.findIndex((bidStrategy) => bidStrategy.uuid === bidStrategyUuid))) {
-          this.bidStrategyUuidSelected = bidStrategyUuid;
-        }
-        this.isLoading = false;
-      });
-
-    this.subscriptions.push(subscription);
   }
 
   onBidStrategySelect(): void {
@@ -97,7 +87,7 @@ export class EditCampaignBidStrategyComponent extends HandleSubscription impleme
         UPDATE_CAMPAIGN_FAILURE
       ),
       first()
-    ).subscribe(action => {
+    ).subscribe((action: Action) => {
       this.submitted = false;
       if (UPDATE_CAMPAIGN_SUCCESS === action.type) {
         this.store.dispatch(new ShowSuccessSnackbar(SAVE_SUCCESS));