+++ src/app/auth/login/login.component.ts
@@ -14,8 +14,10 @@ import { AppState } from 'models/app-state.model'
 import * as authActions from 'store/auth/auth.actions'
 import { Info } from 'models/info.model'
 import AdsWallet from '@adshares/ads-connector'
+import { ADSHARES_WALLET, METAMASK_WALLET } from 'models/enum/link.enum'
 import { WalletToken } from 'models/settings.model'
 import { stringToHex } from 'web3-utils'
+import { ErrorResponseDialogComponent } from 'common/dialog/error-response-dialog/error-response-dialog.component'
 
 @Component({
   selector: 'app-login',
@@ -23,6 +25,8 @@ import { stringToHex } from 'web3-utils'
   styleUrls: ['./login.component.scss'],
 })
 export class LoginComponent extends HandleSubscription implements OnInit {
+  readonly ADSHARES_WALLET = ADSHARES_WALLET
+  readonly METAMASK_WALLET = METAMASK_WALLET
   registrationMode: string
   loginForm: FormGroup
 
@@ -30,8 +34,8 @@ export class LoginComponent extends HandleSubscription implements OnInit {
   loginFormSubmitted: boolean = false
   criteriaError = false
 
-  adsWalletAvailable: boolean = false
-  ethereumAvailable: boolean = false
+  adsWalletAvailable: boolean = true
+  ethereumAvailable: boolean = true
   isAdsLoggingIn: boolean = false
   isBscLoggingIn: boolean = false
   walletLoginError: string | null
@@ -66,10 +70,6 @@ export class LoginComponent extends HandleSubscription implements OnInit {
     this.checkIfUserRemembered()
     this.storeReferralTokenIfPresent()
     this.store.dispatch(new authActions.UserLogOutSuccess())
-
-    const adsWallet = new AdsWallet()
-    adsWallet.getInfo().then(() => (this.adsWalletAvailable = true))
-    this.ethereumAvailable = typeof (window as any).ethereum !== 'undefined'
   }
 
   createForm (): void {
@@ -106,7 +106,17 @@ export class LoginComponent extends HandleSubscription implements OnInit {
       (user: User) => {
         this.processLogin(user)
       },
-      () => {
+      (res) => {
+        if(res.status === 403){
+          this.dialog.open(ErrorResponseDialogComponent, {
+            data: {
+              title: 'Your account is banned',
+              message: `Info: ${res.error.reason } \n\n In case of doubts, please contact support ${appSettings.SUPPORT_EMAIL}`,
+            }
+          })
+          this.isLoggingIn = false
+          return
+        }
         this.criteriaError = true
         this.isLoggingIn = false
       })
@@ -272,10 +282,19 @@ export class LoginComponent extends HandleSubscription implements OnInit {
       }
       this.walletLogin('ADS', response.account.address, token.token, response.signature)
     })
+      .catch(() => {
+        this.adsWalletAvailable = false
+        this.setWalletLoginStatus(false, 'ADS', 'ADS Wallet not installed')
+      })
   }
 
   walletLoginBsc (token: WalletToken): void {
     const ethereum = (window as any).ethereum
+    if (ethereum === undefined) {
+      this.ethereumAvailable = false
+      this.setWalletLoginStatus(false, 'BSC', 'Wallet not installed')
+      return
+    }
     ethereum.request({ method: 'eth_requestAccounts' }).then(
       accounts => {
         ethereum.request({