+++ src/app/models/user.model.ts
@@ -37,18 +37,14 @@ interface User {
   referralToken?: string;
   referralRefundEnabled: boolean;
   referralRefundCommission: number;
+  isBanned: boolean;
+  banReason: string | null;
 }
 
 interface LocalStorageUser extends User {
   remember: boolean;
   passwordLength: number;
   expiration: number;
-}
-
-interface UserRoles extends User {
-  admin: string;
-  publisher: string;
-  advertiser: string;
 }
 
 export { UserAdserverWallet, User, LocalStorageUser, ExchangeRate }