+++ src/app/store/admin/admin.effects.ts
@@ -1,6 +1,12 @@
 import { Injectable } from '@angular/core'
 import { Actions, Effect, ofType } from '@ngrx/effects'
 import {
+  BAN_USER,
+  BanUser,
+  BanUserSuccess,
+  DELETE_USER,
+  DeleteUser,
+  DeleteUserSuccess,
   GET_INDEX,
   GET_LICENSE,
   GET_PRIVACY_SETTINGS,
@@ -15,12 +21,15 @@ import {
   GetRejectedDomainsSuccess,
   GetTermsSettingsSuccess,
   LOAD_ADMIN_SETTINGS,
+  LOAD_ADMIN_SITE_OPTIONS,
   LOAD_ADMIN_WALLET,
   LOAD_ADVERTISERS,
   LOAD_PUBLISHERS,
   LOAD_USERS,
   LoadAdminSettingsFailure,
   LoadAdminSettingsSuccess,
+  LoadAdminSiteOptionsFailure,
+  LoadAdminSiteOptionsSuccess,
   LoadAdminWalletFailure,
   LoadAdminWalletSuccess,
   LoadAdvertisers,
@@ -34,11 +43,14 @@ import {
   LoadUsersSuccess,
   REQUEST_GET_INDEX,
   SET_ADMIN_SETTINGS,
+  SET_ADMIN_SITE_OPTIONS,
   SET_PRIVACY_SETTINGS,
   SET_REJECTED_DOMAINS,
   SET_TERMS_SETTINGS,
   SetAdminSettings,
   SetAdminSettingsSuccess,
+  SetAdminSiteOptions,
+  SetAdminSiteOptionsSuccess,
   SetPrivacySettings,
   SetPrivacySettingsFailure,
   SetPrivacySettingsSuccess,
@@ -46,6 +58,9 @@ import {
   SetRejectedDomainsSuccess,
   SetTermsSettings,
   SetTermsSettingsSuccess,
+  UNBAN_USER,
+  UnbanUser,
+  UnbanUserSuccess,
 } from './admin.actions'
 import { ShowDialogOnError, ShowSuccessSnackbar } from '../common/common.actions'
 import { SAVE_SUCCESS } from 'common/utilities/messages'
@@ -55,7 +70,7 @@ import { catchError, debounceTime, delay, filter, map, switchMap, tap } from 'rx
 import { ClickToADSPipe } from 'common/pipes/adshares-token.pipe'
 import { HTTP_NOT_FOUND } from 'common/utilities/codes'
 import { USER_LOG_OUT_SUCCESS } from 'store/auth/auth.actions'
-import { AdminSettingsResponse } from 'models/settings.model'
+import { AdminSettingsResponse, AdminSiteOptionsResponse } from 'models/settings.model'
 
 @Injectable()
 export class AdminEffects {
@@ -204,6 +219,25 @@ export class AdminEffects {
     )
 
   @Effect()
+  loadAdminSiteOptions$ = this.actions$
+    .pipe(
+      ofType(LOAD_ADMIN_SITE_OPTIONS),
+      switchMap(() => this.service.getAdminSiteOptions()
+        .pipe(
+          map((response: AdminSiteOptionsResponse) => {
+            return <AdminSiteOptionsResponse>{
+              ...response,
+            }
+          }),
+          map(options => new LoadAdminSiteOptionsSuccess(options)),
+          catchError(error => {
+            return observableOf(new LoadAdminSiteOptionsFailure(error))
+          })
+        )
+      )
+    )
+
+  @Effect()
   loadAdminWallet$ = this.actions$
     .pipe(
       ofType(LOAD_ADMIN_WALLET),
@@ -227,6 +261,23 @@ export class AdminEffects {
           ]),
           catchError(() => observableOf(new ShowDialogOnError(
             'We weren\'t able to save your settings this time. Please, try again later'
+          )))
+        )
+      )
+    )
+
+  @Effect()
+  saveAdminSiteOptions$ = this.actions$
+    .pipe(
+      ofType<SetAdminSiteOptions>(SET_ADMIN_SITE_OPTIONS),
+      switchMap(action => this.service.setAdminSiteOptions(action.payload)
+        .pipe(
+          switchMap(() => [
+            new SetAdminSiteOptionsSuccess(action.payload),
+            new ShowSuccessSnackbar(SAVE_SUCCESS)
+          ]),
+          catchError(() => observableOf(new ShowDialogOnError(
+            'We weren\'t able to save your site options this time. Please, try again later'
           )))
         )
       )
@@ -338,4 +389,63 @@ export class AdminEffects {
         )
       )
     )
+
+  @Effect()
+  BanUser$ = this.actions$
+    .pipe(
+      ofType<BanUser>(BAN_USER),
+      switchMap(action => {
+          return this.service.banUser(action.payload)
+            .pipe(
+              switchMap((userInfo) => [
+                new BanUserSuccess(userInfo),
+                new ShowSuccessSnackbar(SAVE_SUCCESS)
+              ]),
+              catchError((response) => AdminEffects.showDialogOnErrorObservable(response))
+            )
+        }
+      )
+    )
+
+  @Effect()
+  UnbanUser$ = this.actions$
+    .pipe(
+      ofType<UnbanUser>(UNBAN_USER),
+      switchMap(action => {
+          return this.service.unbanUser(action.payload)
+            .pipe(
+              switchMap((userInfo) => [
+                new UnbanUserSuccess(userInfo),
+                new ShowSuccessSnackbar(SAVE_SUCCESS)
+              ]),
+              catchError((response) => AdminEffects.showDialogOnErrorObservable(response))
+            )
+        }
+      )
+    )
+
+  @Effect()
+  DeleteUser$ = this.actions$
+    .pipe(
+      ofType<DeleteUser>(DELETE_USER),
+      switchMap(action => {
+          return this.service.deleteUser(action.payload)
+            .pipe(
+              switchMap(() => {
+                return [
+                  new DeleteUserSuccess(action.payload),
+                  new ShowSuccessSnackbar(SAVE_SUCCESS)
+                ]
+              }),
+              catchError((response) => AdminEffects.showDialogOnErrorObservable(response))
+            )
+        }
+      )
+    )
+
+  private static showDialogOnErrorObservable (response) {
+    return observableOf(new ShowDialogOnError(
+      `Operation failure. ${response.error.message}`
+    ))
+  }
 }